Fixed issues with data plotting using ViewModel and LiveData with an Observer which automatically updates the view when data is updated. Further testing required to ensure functionality is as expected over multiple weeks / months / years. Other things to be fixed / updated include:
- Prevent activity resize when softkeyboard opens.
- Update themes (Currently 'Rose's theme only viable theme)
- Change the location of the Snackbar (Displays when all debt paid off - currently covers bottom action buttons when doing so)
- Fix back navigation (Not as expected after changing themes several times)
- Change horizontal view. Currently overcrowded as is. Make more minimal. Possible only the graph and current debt value.
- Fix undo button (Not updating graph as is)

+++ app/src/main/java/syd/jjj/debtcollector/DebtValue.java
@@ -54,13 +54,13 @@ public class DebtValue {
         return "$" + mDollarValue + "." + mCentValue + " added: " + dateString;
     }
 
-    public float getRawX() {
+    public long getRawX() {
         return mDate.getTime();
     }
 
-    public float getRawY() {
+    public long getRawY() {
         String mDollarCentValue = mDollarValue.concat(mCentValue);
-        return Float.valueOf(mDollarCentValue);
+        return Long.valueOf(mDollarCentValue);
     }
